User can only delete own products in category.

+++ backend/controllers/categoryController.js
@@ -52,10 +52,8 @@ const updateCategory = async (req, res) => {
     const category = await Category.findByIdAndUpdate(categoryId, { name })
 
     if (category) {
-        console.log("category", category)
         for (let i = 0; i < category.product_list.length; i++) {
             const product = await Product.findByIdAndUpdate(category.product_list[i], { category: name })
-            if (product) console.log("product", product)
         }
     }
     res.status(200).json({ category: category, products: await Product.find({ category: name }) })
@@ -68,17 +66,29 @@ const deleteCategory = async (req, res) => {
         res.status(400).json({ msg: "Should provide categoryId" })
     }
 
-    const category = await Category.findById(categoryId)
+    let category = await Category.findById(categoryId)
 
 
     if (category) {
         for (let i = 0; i < category?.product_list.length; i++) {
             const productId = category?.product_list[i];
-            await Product.findByIdAndDelete(productId)
+            const product = await Product.findById(productId)
+            if (product?.owner_id == req.user._id) {
+                await product.delete()
+                await Category.findByIdAndUpdate(categoryId, { $pull: { product_list: { $in: [productId] } } })
+            }
         }
-        category.delete()
+        category = await Category.findById(categoryId)
+        if (category.product_list.length == 0) {
+            res.status(200).json({ msg: `Succesfuly deleted ${category.name} category` })
+            category.delete()
+            return
+        } else {
+            res.status(200).json({ msg: `Succesfuly deleted your products which in ${category.name} category` })
+        }
 
-        res.status(200).json({ msg: "Succesfuly deleted", category: category.name })
+
+
     } else {
         res.status(400).json({ msg: "Couldn't find the category." })
     }
